AbstractGenealogy

Types

Functions

Interface

General

Vertices/Edges

Sequences

Ancestrality

Recombinations

Plotting

General

Vertices/Edges

Sequences

Ancestrality

Plotting

Types

Moonshine.AbstractGenealogy — Type

abstract type AbstractGenealogy <: Graphs.SimpleGraphs.AbstractSimpleGraph{Int64}

Abstract type for genealogies.

Implements the Graphs.jl interface so that subtypes implementing the AbstractGenealogy interface can additionally be treated as regular graphs.

Tree & Arg are subtypes that implement the AbstractGenealogy interface.

Moonshine.EdgesInterval — Type

struct EdgesInterval{T, I, E}

Flexible edge iterators that supports various constraints.

Possible constraints are any combination of the following:

An interval of genetic positions. Any non-ancestral edge is ignored.

A minimum latitude. Any edge under that latitude is ignored. An edge e is considered under a latitude l if latitude(genealogy, dst(e)) < l.

A set of predicates. If any of the predicates evaluate to true for a given edge, that edge is ignored.

Do not construct directly, use edges_interval instead.

Fields

genealogy::Any: Genealogy to iterate over

ωs::Any: Interval to consider

buffer::CheapStack: Edges buffer

visited::UnsafeArrays.UnsafeArray{Bool, 1}: True is associated recombination vertex has been visited previously

min_latitude::Float64: Only consider edges located above this latitude

block_predicates::Array{FunctionWrappers.FunctionWrapper{Bool, Tuple{E}}, 1} where E: Edges for which one of the predicates is true are blocked

Constructors

EdgesInterval(
    genealogy,
    ωs,
    buffer,
    visited,
    min_latitude,
    block_predicates
)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Genealogy.jl:1049.

EdgesInterval(genealogy, ωs, stack, visited)
EdgesInterval(genealogy, ωs, stack, visited, root)
EdgesInterval(
    genealogy,
    ωs,
    stack,
    visited,
    root,
    min_latitude
)
EdgesInterval(
    genealogy,
    ωs,
    stack,
    visited,
    root,
    min_latitude,
    block_predicates
)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Genealogy.jl:1063.

EdgesInterval(genealogy, ωs, store, visited)
EdgesInterval(genealogy, ωs, store, visited, root)
EdgesInterval(
    genealogy,
    ωs,
    store,
    visited,
    root,
    min_latitude,
    block_predicates
)
EdgesInterval(
    genealogy,
    ωs,
    store,
    visited,
    root,
    min_latitude
)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Genealogy.jl:1077.

–Internal–

Functions

Interface

AbstractGenealogy interface. Subtypes must implement every method unless otherwise specified.

General

Moonshine.dens — Function

dens(genealogy; logscale = false)

Likelihood of a genealogy.

Implementation

Only mandatory if copulas are to be fitted on the genealogy.

Methods

dens(arg; logscale)
dens(tree; logscale)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/genealogy_common.jl:48.

Moonshine.describe — Function

describe(genealogy, long = true)

Return a string containing a long or short description of a genealogy. Used internally by pretty printing functions.

Methods

describe()
describe(, long)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Arg/Arg.jl:89.

describe()
describe(, long)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Tree.jl:77.

Moonshine.graph — Function

graph(genealogy)

Return the underlying graph of a genealogy.

Methods

graph(arg)
graph(tree)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/genealogy_common.jl:31.

Vertices/Edges

Moonshine.latitudes — Function

latitudes(genealogy[, vs])

Return the latitudes of (a subset of) the internal vertices of a genealogy.

See also latitude to get the latitude of a single vertex.

Implementation

A default implementation for latitudes(::AbstractGenealogy, ::Any) is available; only latitudes(::T) is required.

Methods

latitudes(genealogy, vs)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Genealogy.jl:81.

latitudes(arg)
latitudes(tree)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/genealogy_common.jl:31.

Moonshine.latitude — Function

latitude(genealogy, v)

Latitude of a vertex.

See also latitudes to get the latitudes of internal vertices stored in a genealogy.

Methods

latitude(arg, v)
latitude(tree, v)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/genealogy_common.jl:42.

Moonshine.leaves — Function

leaves(genealogy)

Return an interable containing the leaves of a genealogy.

See also ivertices for internal vertices and nleaves for the number of leaves.

Implementation

Default implementations assume that the first nleaves(genealogy) vertices are the leaves of the genealogy. If this is the case for your type, you do not need to implement this method.

Moonshine.ivertices — Function

ivertices(genealogy)

Return an interable containing the internal vertices of a genealogy.

See also leaves for leaves and nivertices for the number of leaves and internal vertices.

Implementation

Default implementations assume that the first nleaves(genealogy) vertices are the leaves of the genealogy. If this is the case for your type, you do not need to implement this method.

Moonshine.maxdads — Function

maxdads(genealogy)

Maximum possible number of parents for a vertex. See also maxchildren.

Implementation

Must be a generated function.

Methods

maxdads(_)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Arg/Arg.jl:111.

maxdads(_)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Tree.jl:83.

–Internal–

Moonshine.maxchildren — Function

maxchildren(genealogy)

Maximum possible number of children for a vertex. See also maxdads.

Implementation

Must be a generated function.

Methods

maxchildren(_)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Arg/Arg.jl:113.

maxchildren(_)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Tree.jl:85.

–Internal–

Sequences

Moonshine.positions — Function

positions(genealogy)

Positions of the markers.

Methods

positions(sample)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Sample.jl:166.

positions(arg)
positions(tree)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/genealogy_common.jl:35.

Moonshine.sequences — Function

sequences(genealogy[, e])
sequences(genealogy[, vs])

Sequences of a genealogy. If an edge is specified, return the sequences associated with the vertices incident to that edge. If an iterable of vertices is specified, return the sequences associated with these vertices.

See also sequence to get the sequence associated with a vertex.

Implementation

Custom types only need to implement sequences(::T).

Methods

sequences(genealogy, vs)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Genealogy.jl:148.

sequences(genealogy, e)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Genealogy.jl:150.

sequences(arg)
sequences(tree)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/genealogy_common.jl:31.

Ancestrality

Moonshine.ancestral_intervals — Function

ancestral_intervals(genealogy, x)

Interval for which x is ancestral. Default implementation assumes that anything is ancestral for [0, ∞).

See also ancestral_intervals! for a non-allocating alternative.

Methods

ancestral_intervals(arg, e)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Arg/Arg.jl:188.

ancestral_intervals(arg, v)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Arg/Arg.jl:201.

ancestral_intervals(genealogy, x)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Genealogy.jl:205.

Moonshine.ancestral_intervals! — Function

ancestral_intervals!(ωs, genealogy, x)

Interval for which x is ancestral. Default implementation assumes that anything is ancestral for [0, ∞).

See also ancestral_intervals.

Methods

ancestral_intervals!(ωs, arg, e; wipe, simplify)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Arg/Arg.jl:180.

ancestral_intervals!(ωs, arg, v; wipe)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Arg/Arg.jl:190.

ancestral_intervals!(ωs, _, _)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Genealogy.jl:222.

Moonshine.mrca — Function

mrca(genealogy[, vs = leaves(genealogy), ωs = Ω(0, ∞)])

Most recent common ancestor of a set of vertices.

See also tmrca for the time to the most recent common ancestor.

Methods

mrca(arg)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Arg/Arg.jl:107.

mrca(arg, ωs)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Arg/Arg.jl:109.

mrca(tree)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Tree.jl:79.

mrca(tree, _)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Tree.jl:81.

mrca(tree, vs, x; buffer)
mrca(arg, vs, x; buffer)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/genealogy_common.jl:77.

mrca(tree, vs)
mrca(arg, vs)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/genealogy_common.jl:101.

Recombinations

Moonshine.recombinations — Function

recombinations(genealogy)

Iterator over the recombination vertices of a genealogy.

Implementation

Default implementation returns an empty iterator.

Methods

recombinations(arg)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Arg/Arg.jl:95.

recombinations(_)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Genealogy.jl:259.

Moonshine.isrecombination — Function

isrecombination(genealogy, v)

Returns true if vertex v is a recombination for genealogy.

Default implementation always returns false.

Methods

isrecombination(_, v, n)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Arg/Arg.jl:91.

isrecombination(arg, v)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Arg/Arg.jl:93.

isrecombination(_, _)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Genealogy.jl:274.

Moonshine.nrecombinations — Function

nrecombinations(genealogy)

Number of recombinations.

Implementation

Default implementation returns 0.

Methods

nrecombinations(arg)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Arg/Arg.jl:105.

nrecombinations(_)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Genealogy.jl:243.

Plotting

Moonshine.plot_layout — Function

plot_layout(genealogy)

Layout function for genealogy plotting.

Implementation

Defaults to Spring().

Methods

plot_layout(arg)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Arg/Arg.jl:164.

plot_layout(_)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Genealogy.jl:291.

–Internal–

General

Vertices/Edges

Moonshine.nleaves — Function

nleaves(genealogy)

Number of leaves in a genealogy.

See also leaves for an iterable over leaves and nivertices for an internal vertices counterpart.

Methods

nleaves(arg)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Arg/Arg.jl:87.

nleaves(genealogy)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Genealogy.jl:609.

nleaves(tree)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Tree.jl:75.

Moonshine.nivertices — Function

nivertices(genealogy)

Number of internal vertices in a genealogy.

See also ivertices for an iterable over internal vertices and nleaves for a leaves counterpart.

Moonshine.isleaf — Function

isleaf(genealogy, v)

True if v is a leaf.

See also isivertex, isroot.

Moonshine.isroot — Function

isroot(genealogy, v)

True if v is the root.

See also isleaf, isivertex.

Moonshine.isivertex — Function

isivertex(genealogy, v)

True if v is an internal vertex.

See also isleaf, isroot.

Moonshine.branchlength — Function

branchlength(genealogy)
branchlength(genealogy, ωs)
branchlength(genealogy, e)

Total branch length of a genealogy. If an interval is specified, returns the branch length of the associated marginal genealogy. If an edge is specified, returns the length of that edge.

Methods

branchlength(genealogy, e)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Genealogy.jl:661.

branchlength(genealogy)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Genealogy.jl:664.

branchlength(genealogy, ωs; buffer)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Genealogy.jl:666.

branchlength(genealogy, s, d)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Genealogy.jl:1021.

Moonshine.edges_interval — Function

edges_interval(genealogy, ωs)
edges_interval(genealogy, ωs, buffer, visited, root = mrca(genealogy), min_latitude = zero(Float64), block_predicates = [])

Iterage over a genealogy's edges via EdgesInterval.

buffer can be either a CheapStack or an UnsafeArray that will be used as buffer for a newly constructed CheapStack.

Methods

edges_interval(
    genealogy,
    ωs,
    buffer,
    visited,
    root,
    min_latitude;
    block_predicates
)
edges_interval(genealogy, ωs, buffer, visited, root; ...)
edges_interval(genealogy, ωs, buffer, visited; ...)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Genealogy.jl:1138.

edges_interval(genealogy, ωs)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Genealogy.jl:1143.

Moonshine.edgesmap — Function

edgesmap(genealogy)

Return a Dict that maps every edge of a genealogy to an integer in 1:ne(genealogy).

Moonshine.nlive — Function

nlive(genealogy, lat, ωs; block_predicates, buffer)

Number of live edges in a (marginal) genealogy at a given latitude.

block_predicate is passed directly to edges_interval.

See also nlive!.

Moonshine.nlive! — Function

nlive!(counts, genealogy, lats, ωs[, stack]; block_predicates = [], buffer = default_buffer())

Number of live edges in a (marginal) genealogy at a given latitude.

Counts are stored in counts which is filled with 0s initially. ̀lats must be the same size as counts.

block_predicate and stack are passed directly to edges_interval.

See also nlive.

Methods

nlive!(
    counts,
    genealogy,
    lats,
    ωs,
    stack;
    block_predicates,
    buffer
)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Genealogy.jl:1209.

nlive!(
    counts,
    genealogy,
    lats,
    ωs;
    block_predicates,
    buffer
)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Genealogy.jl:1238.

Moonshine.ismutation_edge — Function

ismutation_edge(genealogy, e, idx)

Determines if an edge is a mutation edge for a given marker.

Sequences

Moonshine.sequence — Function

sequence(genealogy, v)

Sequence of a genealogy associated with a vertex.

See also sequences to get all the sequences of a genealogy.

Moonshine.idxtopos — Function

idxtopos(genealogy, idx)

Return the position of the marker given its index.

Methods

idxtopos(genealogy, idx)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Genealogy.jl:351.

idxtopos(sample, idx)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Sample.jl:172.

Moonshine.postoidx — Function

postoidx(genealogy, pos)

Return the largest marker's index that is at a position lesser than the one given.

Methods

postoidx(genealogy, pos)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Genealogy.jl:351.

postoidx(sample, pos)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Sample.jl:180.

postoidx(sample, ω)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Sample.jl:223.

Moonshine.nmarkers — Function

nmarkers(genealogy[, ωs])

Number of markers in the sequences of a genealogy. If an interval ωs is specified, returns the number of markers contained in that interval.

Methods

nmarkers(genealogy, ωs)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Genealogy.jl:644.

nmarkers(sample)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Sample.jl:168.

nmarkers(sample, ω)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Sample.jl:170.

Moonshine.nmutations — Function

nmutations(genealogy[, e])

Number of mutation on a genealogy. If an edge is specified, return the number of mutations on that edge.

See also nmutations! for a non-allocating alternative.

Methods

nmutations(genealogy)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Genealogy.jl:1008.

nmutations(genealogy, e)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Genealogy.jl:1014.

nmutations(genealogy, s, d)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Genealogy.jl:1021.

Moonshine.nmutations! — Function

nmutations!(mask, genealogy, e)

Number of mutation on a genealogy. If an edge is specified, return the number of mutations on that edge.

See also nmutations.

Ancestrality

Moonshine.dad — Function

dad(genealogy, v)

Return the parent of a vertex or 0 if none. It only makes sense to use this method if you know v has a single parent. Otherwise use dads.

See also child, descendants, ancestors and ̀siblings.

Moonshine.dads — Function

dads(genealogy, v[, ωs])

Parents of a vertex, optionally restricted to a marginal genealogy. If you know in advance that v has a single dad, use ̀̀dad instead.

The following rules are used to decide if an edge e is ancestral:

If ωs is a number, the ancestral interval of e must cover ωs.

If ωs is an Ω or a set of Ωs, the intersection of the ancestral interval of e with ωs must be non-empty.

Danger

Return a reference to the underlying adjacency lists. No touchy!

See also child, children, sibling, siblings, descendants and ancestors.

Methods

dads(arg, v, ωs)
dads(arg, v, ωs)
dads(arg, v, ωs)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Arg/Arg.jl:123.

dads(genealogy, v)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Genealogy.jl:746.

Moonshine.child — Function

child(genealogy, v)

Return the child of a vertex or 0 if none. It only makes sense to use this method if you know v has a single child. Otherwise use children.

See also dad, ancestors, descendants and ̀siblings.

Moonshine.children — Function

children(genealogy, v[, ωs])

Children of a vertex, optionally restricted to a marginal genealogy. If you know in advance that v has a single child, use ̀̀child instead.

The following rules are used to decide if an edge e is ancestral:

If ωs is a number, the ancestral interval of e must cover ωs.

If ωs is an Ω or a set of Ωs, the intersection of the ancestral interval of e with ωs must be non-empty.

Danger

Return a reference to the underlying adjacency lists. No touchy!

See also dad, dads, sibling, siblings, descendants and ancestors.

Methods

children(arg, v, ωs)
children(arg, v, ωs)
children(arg, v, ωs)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Arg/Arg.jl:123.

children(genealogy, v)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Genealogy.jl:746.

Moonshine.ancestors — Function

ancestors(genealogy, v[, ωs])
ancestors(genealogy, v[, ωs], buf_ptr)

Ancestors of a vertex, optionally restricted to a marginal genealogy.

A pointer to some kind of buffer (an array for instance) can be provided to avoid allocation. In that case, an UnsafeArray wrapped around it will be returned.

The following rules are used to decide if an edge e is ancestral:

If ωs is a number, the ancestral interval of e must cover ωs.

If ωs is an Ω or a set of Ωs, the intersection of the ancestral interval of e with ωs must be non-empty.

See also child, dad, children, dads, sibling, siblings and descendants.

Methods

ancestors(genealogy, v)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Genealogy.jl:823.

ancestors(genealogy, v, ω)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Genealogy.jl:856.

Moonshine.descendants — Function

descendants(genealogy, v[, ωs])
descendants(genealogy, v[, ωs], buf_ptr)

Descendants of a vertex, optionally restricted to a marginal genealogy.

A pointer to some kind of buffer (an array for instance) can be provided to avoid allocation. In that case, an UnsafeArray wrapped around it will be returned.

The following rules are used to decide if an edge e is ancestral:

If ωs is a number, the ancestral interval of e must cover ωs.

If ωs is an Ω or a set of Ωs, the intersection of the ancestral interval of e with ωs must be non-empty.

See also child, dad, children, dads, sibling, siblings and ancestors.

Methods

descendants(genealogy, v)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Genealogy.jl:823.

descendants(genealogy, v, ω)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Genealogy.jl:856.

Moonshine.sibling — Function

sibling(genealogy, v)

Sibling of a vertex, that is the other vertex in the genealogy that have the same parent. It only makes sense to use this method if you know v has a single sibling. Otherwise use siblings.

See also child, dad, children, dads, descendants and ancestors.

Moonshine.siblings — Function

siblings(genealogy, v[, ωs])
siblings(genealogy, v[, ωs], buf_ptr)

Siblings of a vertex, that is the other vertices in the genealogy that share at least one parent, optionally restricted to a marginal genealogy.

A pointer to some kind of buffer (an array for instance) can be provided to avoid allocation. In that case, an UnsafeArray wrapped around it will be returned.

If you know in advance that v has a single sibling, you can use sibling instead.

See also child, dad, children, dads, descendants and ancestors.

Methods

siblings(genealogy, v, args)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Genealogy.jl:914.

Moonshine.ancestral_mask — Function

ancestral_mask(reference, x; ωs_buf = Set{Ω}())

Mask non ancestral positions to 0. If wipe = true, all markers in η will be initialized to 0.

Methods

ancestral_mask(arg, x)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Arg/Arg.jl:228.

ancestral_mask(e, arg)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Arg/Arg.jl:233.

ancestral_mask(genealogy, x)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Genealogy.jl:366.

ancestral_mask(sample, x)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Sample.jl:329.

Moonshine.ancestral_mask! — Function

ancestral_mask!(η, reference, x; ωs_buf = Set{Ω}(), wipe = true)

Mask non ancestral positions to 0. If wipe = true, all markers in η will be initialized to 0.

Methods

ancestral_mask!(η, arg, e; wipe)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Arg/Arg.jl:203.

ancestral_mask!(h, arg, v; buffer, wipe)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Arg/Arg.jl:206.

ancestral_mask!(η, genealogy, x; wipe)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Genealogy.jl:382.

ancestral_mask!(v, sample, ω; wipe)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Sample.jl:303.

ancestral_mask!(η, sample, ω; wipe)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Sample.jl:313.

ancestral_mask!(η, sample, ωs; wipe)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Sample.jl:319.

ancestral_mask!(η, sample, pos; wipe)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Sample.jl:332.

Moonshine.tmrca — Function

tmrca(genealogy[, vs])

Time to the most recent common ancestor (MRCA) of (a subset of) vertices.

Methods

tmrca(genealogy)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Genealogy.jl:688.

tmrca(genealogy, vs)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Genealogy.jl:695.

tmrca(arg, vs, ωs)
tmrca(tree, vs, ωs)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/genealogy_common.jl:103.

Plotting

GraphMakie.graphplot — Method

graphplot(graph::AbstractGraph)
graphplot!(ax, graph::AbstractGraph)

Creates a plot of the network graph. Consists of multiple steps:

Layout the nodes: see layout attribute. The node position is accessible from outside the plot object p as an observable using p[:node_pos].

plot edges as edgeplot-plot

if arrow_show plot arrowheads as scatter-plot

plot nodes as scatter-plot

if nlabels!=nothing plot node labels as text-plot

if elabels!=nothing plot edge labels as text-plot

The main attributes for the subplots are exposed as attributes for graphplot. Additional attributes for the scatter, edgeplot and text plots can be provided as a named tuples to node_attr, edge_attr, nlabels_attr and elabels_attr.

Most of the arguments can be either given as a vector of length of the edges/nodes or as a single value. One might run into errors when changing the underlying graph and therefore changing the number of Edges/Nodes.

Attributes

Main attributes

layout=Spring(): function AbstractGraph->Vector{Point} or Vector{Point} that determines the base layout. Can also be any network layout from NetworkLayout.jl, like Spring, Stress, Spectral, etc.

node_color=automatic: Defaults to scatter_theme.color in absence of ilabels.

node_size=automatic: Defaults to scatter_theme.markersize in absence of ilabels. Otherwise choses node size based on ilabels size.

node_marker=automatic: Defaults to scatter_theme.marker in absence of ilabels.

node_strokewidth=automatic Defaults to scatter_theme.strokewidth in absence of ilabels.

node_attr=(;): List of kw arguments which gets passed to the scatter command

edge_color=lineseg_theme.color: Color for edges.

edge_width=lineseg_theme.linewidth: Pass a vector with 2 width per edge to get pointy edges.

edge_attr=(;): List of kw arguments which gets passed to the linesegments command

arrow_show=Makie.automatic: Bool, indicate edge directions with arrowheads? Defaults to Graphs.is_directed(graph).

arrow_marker='➤'

arrow_size=scatter_theme.markersize: Size of arrowheads.

arrow_shift=0.5: Shift arrow position from source (0) to dest (1) node. If arrow_shift=:end, the arrowhead will be placed on the surface of the destination node (assuming the destination node is circular).

arrow_attr=(;): List of kw arguments which gets passed to the scatter command

Node labels

The position of each label is determined by the node position plus an offset in data space.

nlabels=nothing: Vector{String} with label for each node

nlabels_align=(:left, :bottom): Anchor of text field.

nlabels_distance=0.0: Pixel distance from node in direction of align.

nlabels_color=labels_theme.color

nlabels_offset=nothing: Point or Vector{Point} (in data space)

nlabels_fontsize=labels_theme.fontsize

nlabels_attr=(;): List of kw arguments which gets passed to the text command

Inner node labels

Put labels inside the marker. If labels are provided, change default attributes to node_marker=Circle, node_strokewidth=1 and node_color=:gray80. The node_size will match size of the ilabels.

ilabels=nothing: Vector with label for each node

ilabels_color=labels_theme.color

ilabels_fontsize=labels_theme.fontsize

ilabels_attr=(;): List of kw arguments which gets passed to the text command

Edge labels

The base position of each label is determined by src + shift*(dst-src). The additional distance parameter is given in pixels and shifts the text away from the edge.

elabels=nothing: Vector{String} with label for each edge

elabels_align=(:center, :center): Anchor of text field.

elabels_side = :left: Side of the edge to put the edge label text

elabels_distance=Makie.automatic: Pixel distance of anchor to edge. The direction is decided based on elabels_side

elabels_shift=0.5: Position between src and dst of edge.

elabels_rotation=Makie.automatic: Angle of text per label. If nothing this will be determined by the edge angle. If automatic it will also point upwards making it easy to read.

elabels_offset=nothing: Additional offset in data space

elabels_color=labels_theme.color

elabels_fontsize=labels_theme.fontsize

elabels_attr=(;): List of kw arguments which gets passed to the text command

Curvy edges & self edges/loops

edge_plottype=Makie.automatic(): Either automatic, :linesegments or :beziersegments. :beziersegments are much slower for big graphs!

Self edges / loops:

selfedge_size=Makie.automatic(): Size of selfloop (dict/vector possible).

selfedge_direction=Makie.automatic(): Direction of center of the selfloop as Point2 (dict/vector possible).

selfedge_width=Makie.automatic(): Opening of selfloop in rad (dict/vector possible).

Note: If valid waypoints are provided for selfloops, the selfedge attributes above will be ignored.

High level interface for curvy edges:

curve_distance=0.1:

Specify a distance of the (now curved) line to the straight line in data space. Can be single value, array or dict. User provided tangents or waypoints will overrule this property.

curve_distance_usage=Makie.automatic():

If Makie.automatic(), only plot double edges in a curvy way. Other options are true and false.

Tangents interface for curvy edges:

tangents=nothing:

Specify a pair of tangent vectors per edge (for src and dst). If nothing (or edge idx not in dict) draw a straight line.

tfactor=0.6:

Factor is used to calculate the bezier waypoints from the (normalized) tangents. Higher factor means bigger radius. Can be tuple per edge to specify different factor for src and dst.

Note: Tangents are ignored on selfloops if no waypoints are provided.

Waypoints along edges:

waypoints=nothing

Specify waypoints for edges. This parameter should be given as a vector or dict. Waypoints will be crossed using natural cubic splines. The waypoints may or may not include the src/dst positions.

waypoint_radius=nothing

If the attribute waypoint_radius is nothing or :spline the waypoints will be crossed using natural cubic spline interpolation. If number (dict/vector possible), the waypoints won't be reached, instead they will be connected with straight lines which bend in the given radius around the waypoints.

Moonshine.plot_latitudes — Function

plot_latitudes(genealogy; kwargs...)

Unicode histogram of a genealogy's latitudes. Additional keywords arguments are passed directly to UnicodePlots.histogram.

See also latitudes, latitude