Sample

Moonshine.Sample — Type

struct Sample <: AbstractVector{Sequence}

Contain a sample of haplotypes and informations about them.

Implements the iteration interface and the array interface.

Fields

H::Vector{Sequence}: Vector of haplotypes

μ::Float64: Unscaled (per-locus) mutation rate

ρ::Float64: Unscaled (per-locus) recombination rate

Ne::Float64: Effective population size

sequence_length::Float64: Sequence length

positions::Vector{Float64}: Marker's positions

coefs::Tuple{Float64, Float64}: Coefficients for positions' line

Constructors

Info

Random constructor sample sequences via msprime using a binary mutation model.

Sample(H, μ, ρ, Ne, sequence_length, positions, coefs)
Sample(H, μ, ρ, Ne, sequence_length, positions, coefs)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Sample.jl:40.

Sample(H; μ, ρ, Ne, positions, sequence_length)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Sample.jl:56.

Sample(ts)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Sample.jl:71.

Sample(rng, n, μ, ρ, Ne, sequence_length)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Sample.jl:109.

Sample(mat, positions; μ, ρ, Ne)

defined at /home/runner/work/Moonshine.jl/Moonshine.jl/src/Sample.jl:134.

where:

H: haplotypes

μ (-1): mutation rate

ρ (-1): recombination rate

Ne (1): effective population size

positions: position of markers

sequence_length (maximum(positions)): sequence length

n: number of sequences

rng: random number generator

ts: Tree Sequence

mat: matrix of haplotypes (haplotypes -> columns)

Moonshine.mut_rate — Function

mut_rate(sample, scaled = true)

(Scaled) mutation rate.

Moonshine.rec_rate — Function

rec_rate(sample, scaled = true)

(Scaled) recombination rate.