Global Values

Moonshine.VertexType — Type

const VertexType = Int32

Type of the vertices. Any genealogy that is an instance of AbstractGenealogy should be constructed so that graph(genealogy) is an instance of AbstractGraphs{VertexType} (see Graphs.AbstractGraph.)

–Internal–

Moonshine.AI — Type

const AI = AbstractInterval

Abstract types for intervals from IntervalSets.

See also IntervalSets.AbstractInterval and Ω.

–Internal–

Moonshine.mmn_chunksize — Constant

const mmn_chunksize = 1

Size (in bytes) of a chunk of markers in next_inconsistent_idx.

–Internal–

Moonshine.mmn_chunktype — Type

const mmn_chunktype = (eval ∘ Symbol)("UInt" * string(8mmn_chunksize))

Type of a chunk of markers in next_inconsistent_idx.

–Internal–

Moonshine.simd_vecsize — Constant

const simd_vecsize = simdbytes()

Size (in bytes) of a SIMD registry on host machine.

–Internal–

Moonshine.simd_chunksize — Constant

const simd_chunksize = simd_vecsize ÷ 8

Number of chunks of markers that fit in a SIMD registry.

–Internal–

Moonshine.default_colormap — Constant

default_colormap = :Paired_3

Color map used by UnicodePlots.jl. See ColorSchemes.jl.

Moonshine.clat_gridsize — Constant

const clat_gridsize = 25

Size of the grid for the quadrature step in the recoalescence latitude sampling. procedure.

Customizable via Preferences.

–Internal–